// src/app/models/usuario.ts
export interface Usuario {
  id: number;
  name: string;
  nom_fis: string;
  nom_com: string;
  emp: string;
  emp_div: string;
  cif: string;
  tlf: string;
  eml: string;
  dir: string;
  pai: number;
  pro: number;
  ent_rel: number;
  ent_rel_name: string;
  nom_ent_rel: string;
  usr_web: string;
  pwd_web: string;
  off: boolean;
  pgc_clt: string;
  aux_clt: string;
  num_plan_adq: number;
}